+++ manual-tests/test-release-singlethreaded.ts
@@ -1,26 +1,8 @@
-import { Chessalyzer } from 'chessalyzer.js';
+import { analyzePGN } from 'chessalyzer.js';
 
-import CustomGameTracker from './custom-game-tracker.ts';
-
-// const a = new Tracker.Game();
-// const b = new Tracker.Piece();
-// const c = new Tracker.Tile();
-// const d = new CustomGameTracker();
-
-// a.cfg.profilingActive = true;
-// b.cfg.profilingActive = true;
-// c.cfg.profilingActive = true;
-
-const header = await Chessalyzer.analyzePGN(
-    './pgn/lichess_db_standard_rated_2013-12.pgn',
-    undefined,
-    null,
-);
+console.time('Time elapsed');
+const header = await analyzePGN('./pgn/lichess_db_standard_rated_2013-12.pgn', {
+    workers: false,
+});
 console.log(header);
-console.log(
-    `${header.cntGames} games (${header.cntMoves} moves) processed (${header.mps} moves/s)`,
-);
-// console.log('Game: ' + Math.round(a.time) / 1000);
-// console.log('Piece: ' + Math.round(b.time) / 1000);
-// console.log('Tile: ' + Math.round(c.time) / 1000);
-// console.log('Custom: ' + d.wins);
+console.timeEnd('Time elapsed');