+++ src/index.ts
@@ -1,6 +1,16 @@
-export { default as Chessalyzer } from '#core/chessalyzer';
+export { analyzePGN, printHeatmap } from '#core/analyze';
+export { GameTrackerBase, MoveTracker } from '#tracker/base-tracker';
+export { default as BaseTracker } from '#tracker/base-tracker';
 export { default as PieceTracker, isTrackedPiece } from '#tracker/piece-tracker-base';
 export { default as GameTracker } from '#tracker/game-tracker-base';
 export { default as TileTracker } from '#tracker/tile/tile-tracker-base';
-export { default as BaseTracker } from '#tracker/base-tracker';
-export type { GameAndMoveCountFull } from '#types/analysis';
+export type { Action } from '#types/actions';
+export type {
+    AnalyzeOptions,
+    AnalyzeResult,
+    AnalyzeRun,
+    AnalyzeRunResult,
+    WorkerOptions,
+} from '#types/analysis';
+export type { Game } from '#types/game';
+export type { HeatmapData, Tracker } from '#types/tracker';